Helyesen működő backend, még a fájl mentést meg kell valósítani

+++ backend/src/main/java/hu/szakdolgozat/backend/ocr/OcrResult.java
@@ -39,7 +39,7 @@ public class OcrResult
 		this.ocrResultFile = ocrResultFile;
 	}
 	
-	public OcrResult(String projectName, Integer numberOfSentence, Integer numberOfWords, Integer averageWordCount, File ocrResultFile)
+	public OcrResult(String projectName, Integer numberOfSentence, Integer numberOfWords, double averageWordCount, File ocrResultFile)
 	{
 		this.projectName = projectName;
 		this.numberOfSentence = numberOfSentence;
